Renamed main layout filed and adjusted imports.
