+++ src/features/employees/index.tsx
@@ -4,7 +4,7 @@ import { useAppSelector } from 'store/hooks';
 import { selectCurrentUser } from 'store/slices/authSlice';
 import { useGetEmployeesQuery } from 'store/slices/employeesApiSlice';
 import LoadingSpinner from 'components/utils/LoadingSpinner';
-import MainLayout from 'components/layout/MainLayout';
+import MainLayout from 'components/layout';
 import Navbar from 'components/navigation/NavBar';
 import Pagination from 'components/pagination';
 import EmployeesTable from 'features/employees/EmployeesTable';
