+++ src/features/projectReporting/index.tsx
@@ -1,4 +1,4 @@
-import MainLayout from 'components/layout/MainLayout';
+import MainLayout from 'components/layout';
 
 const ProjectReporting = () => {
 	return (
