+++ src/features/projects/index.tsx
@@ -4,7 +4,7 @@ import { useAppSelector } from 'store/hooks';
 import { selectCurrentUser } from 'store/slices/authSlice';
 import { useGetProjectsQuery } from 'store/slices/projectsApiSlice';
 import LoadingSpinner from 'components/utils/LoadingSpinner';
-import MainLayout from 'components/layout/MainLayout';
+import MainLayout from 'components/layout';
 import Navbar from 'components/navigation/NavBar';
 import Pagination from 'components/pagination';
 import ProjectsTable from 'features/projects/ProjectsTable';
